+++ components/popconfirm/PurePanel.tsx
@@ -6,7 +6,7 @@ import type { PopconfirmProps } from '.';
 import ActionButton from '../_util/ActionButton';
 import { getRenderPropValue } from '../_util/getRenderPropValue';
 import type { SemanticClassNames, SemanticStyles } from '../_util/hooks';
-import Button from '../button';
+import Button from '../button/Button';
 import { convertLegacyProps } from '../button/buttonHelpers';
 import { ConfigContext } from '../config-provider';
 import { useLocale } from '../locale';
