+++ components/upload/UploadList/index.tsx
@@ -11,8 +11,8 @@ import { clsx } from 'clsx';
 import { useForceUpdate } from '../../_util/hooks';
 import initCollapseMotion from '../../_util/motion';
 import { cloneElement } from '../../_util/reactNode';
-import type { ButtonProps } from '../../button';
-import Button from '../../button';
+import type { ButtonProps } from '../../button/Button';
+import Button from '../../button/Button';
 import { ConfigContext } from '../../config-provider';
 import type { UploadFile, UploadListProps } from '../interface';
 import { isImageUrl, previewImage } from '../utils';
